refactor(frontend): split api.ts into domain modules under src/api/
