+++ src/api/invoke.ts
@@ -1,0 +1,1 @@
+export { invoke } from "@tauri-apps/api/core";
